Convert entire transaction service from BigDecimal to Double for simplicity
- Updated Transaction entity to use Double instead of BigDecimal for amount field
- Removed precision/scale annotations that aren't valid for floating point numbers
- Updated TransactionResponse DTO to use Double
- Updated AccountsServiceClient to return/accept Double values
- Updated TransactionService to use simple arithmetic operations instead of BigDecimal methods
- Fixed all test files to use Double values throughout
- Removed BigDecimal imports from all files

This simplifies the banking system to use regular Double values like most modern banking applications, while maintaining proper JSON serialization and database storage.

+++ backend/transactions-service/src/main/java/com/banking/transactionsservice/dto/TransactionResponse.java
@@ -2,7 +2,6 @@ package com.banking.transactionsservice.dto;
 
 import com.banking.transactionsservice.entity.Transaction;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class TransactionResponse {
@@ -11,7 +10,7 @@ public class TransactionResponse {
     private Long userId;
     private Long fromAccountId;
     private Long toAccountId;
-    private BigDecimal amount;
+    private Double amount;
     private Transaction.TransactionType type;
     private String description;
     private Transaction.TransactionStatus status;
@@ -20,7 +19,7 @@ public class TransactionResponse {
     
     public TransactionResponse() {}
     
-    public TransactionResponse(Long id, Long userId, Long fromAccountId, Long toAccountId, BigDecimal amount,
+    public TransactionResponse(Long id, Long userId, Long fromAccountId, Long toAccountId, Double amount,
                               Transaction.TransactionType type, String description, 
                               Transaction.TransactionStatus status, LocalDateTime createdAt, 
                               LocalDateTime processedAt) {
@@ -60,11 +59,11 @@ public class TransactionResponse {
         this.toAccountId = toAccountId;
     }
     
-    public BigDecimal getAmount() {
+    public Double getAmount() {
         return amount;
     }
     
-    public void setAmount(BigDecimal amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
